requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
simple-lama-inpainting>=0.1.0
Pillow>=9.0.0
opencv-python>=4.7.0
numpy>=1.24.0
click>=8.0.0
